moved layer to tf_util
